Remove unnecessary casts

+++ TeraFinder.Plugins/Utils/ImagesUtil.cs
@@ -11,14 +11,14 @@ public static class ImagesUtil
     {
         SpriteName.AllowShinySprite = true;
         var file = pkm.Shiny > TeraShiny.No && pkm.Species < (ushort)Species.Sprigatito ?
-            'b' + SpriteName.GetResourceStringSprite(pkm.Species, pkm.Form, (byte)(int)pkm.Gender, (uint)(pkm.Species == (ushort)Species.Gholdengo ? 999 : 0), EntityContext.Gen9, true) :
-            'a' + SpriteName.GetResourceStringSprite(pkm.Species, pkm.Form, (byte)(int)pkm.Gender, (uint)(pkm.Species == (ushort)Species.Gholdengo ? 999 : 0), EntityContext.Gen9, false);
+            'b' + SpriteName.GetResourceStringSprite(pkm.Species, pkm.Form, (byte)pkm.Gender, (uint)(pkm.Species == (ushort)Species.Gholdengo ? 999 : 0), EntityContext.Gen9, true) :
+            'a' + SpriteName.GetResourceStringSprite(pkm.Species, pkm.Form, (byte)pkm.Gender, (uint)(pkm.Species == (ushort)Species.Gholdengo ? 999 : 0), EntityContext.Gen9, false);
 
         var sprite = (Image?)PKHeX.Drawing.PokeSprite.Properties.Resources.ResourceManager.GetObject(file);
 
         if (sprite is null)
         {
-            file = 'a' + SpriteName.GetResourceStringSprite(pkm.Species, pkm.Form, (byte)(int)pkm.Gender, (uint)(pkm.Species == (ushort)Species.Gholdengo ? 999 : 0), EntityContext.Gen9, false);
+            file = 'a' + SpriteName.GetResourceStringSprite(pkm.Species, pkm.Form, (byte)pkm.Gender, (uint)(pkm.Species == (ushort)Species.Gholdengo ? 999 : 0), EntityContext.Gen9, false);
             sprite = (Image?)PKHeX.Drawing.PokeSprite.Properties.Resources.ResourceManager.GetObject(file);
         }
